# Conflicts:
#	src/main/java/com/springCoupon/Repositories/CouponRepository.java
#	src/main/java/com/springCoupon/Services/AdminService.java
#	src/main/java/com/springCoupon/Services/CompanyService.java
#	src/main/java/com/springCoupon/Services/CustomerService.java
#	src/main/java/com/springCoupon/SpringCouponApplication.java
#	src/main/java/com/springCoupon/controllers/AdminController.java
#	src/main/java/com/springCoupon/controllers/CompanyController.java
#	src/main/java/com/springCoupon/controllers/CustomerController.java
#	src/main/java/com/springCoupon/menus/Admin.java
#	src/main/java/com/springCoupon/menus/LoginManager.java
#	src/main/java/com/springCoupon/menus/SpringMain.java
#	src/main/java/com/springCoupon/menus/prev-test-main
#	src/main/java/com/springCoupon/utilities/DailyJob.java
#	src/main/java/com/springCoupon/utilities/LoginInfo.java
#	src/main/java/com/springCoupon/utilities/TokensList.java

+++ src/main/java/com/springCoupon/utilities/DailyJob.java
@@ -1,52 +1,40 @@
 package com.springCoupon.utilities;
 
-
 import com.springCoupon.Entities.Coupon;
-import com.springCoupon.Entities.Customer;
 import com.springCoupon.Repositories.CouponRepository;
-import com.springCoupon.Repositories.CustomerRepository;
-import com.springCoupon.Services.MainService;
-;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
+import org.springframework.scheduling.annotation.EnableScheduling;
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Component;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 import java.util.stream.Collectors;
 
-@Service
+@Component
+//@EnableScheduling
 public class DailyJob {
-    // annotation @scheduler - spring boot to che
+    // annotation @scheduler - spring boot to check
 
     @Autowired
     CouponRepository couponRepository;
 
-    @Autowired
-    CustomerRepository customerRepository;
-
-    int timeToSleep = 24 * 60 * 60;
-
-
-    Thread thread = new Thread(() -> {
-        List<Coupon> couponList = couponRepository.findByEndDateLessThan(LocalDateTime.now());
-
-        for (Coupon coupon : couponList) {
-        couponRepository.deleteCoupon(coupon.getCouponId());
-        }
-        try {
-            Thread.sleep(timeToSleep);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-    });
-
     public void startDailyJob() {
-
-        thread.start();
-
+        dailyJob();
     }
 
+//    @Scheduled(fixedDelay = 24 * 60 * 60)
+    public void dailyJob() {
+
+        List<Coupon> couponsToDelete = couponRepository.findByEndDateLessThan(LocalDateTime.now());
+        if (!couponsToDelete.isEmpty()) {
+            couponsToDelete.forEach(coupon -> {
+                couponRepository.deletePurchasesOfCustomer(coupon.getCouponId());
+                couponRepository.deleteCoupon(coupon.getCouponId());
+
+            });
+        }
+
+    }
 }
